fix: use and export ValidationError
