+++ packages/helix-shared-config/src/ConfigValidator.js
@@ -14,6 +14,7 @@
 
 const Ajv = require('ajv').default;
 const ajvFormats = require('ajv-formats');
+const ValidationError = require('./ValidationError.js');
 
 const schemas = [
   /* eslint-disable global-require */
@@ -30,43 +31,6 @@ const schemas = [
   require('./schemas/conditions.schema.json'),
   /* eslint-enable global-require */
 ];
-
-class ValidationError extends Error {
-  constructor(msg, errors = []) {
-    function prettyname(path, schema) {
-      if (path.startsWith('.strains')) {
-        return `${schema.title || 'Invalid Strain'} ${path.replace(/\.strains(\.|\[')(.*)/, '$2').replace(/'.*/, '')}`;
-      }
-      return `${schema.title || schema.$id} ${path}`;
-    }
-
-    const detail = errors.map(({
-      keyword, dataPath, message, data, params, parentSchema,
-    }) => {
-      if (keyword === 'additionalProperties') {
-        return `${prettyname(dataPath, parentSchema)} has unknown property '${params.additionalProperty}'`;
-      }
-      if (keyword === 'required' && dataPath === '') {
-        return 'A set of strains and a default strain are missing.';
-      }
-      if (keyword === 'required' && dataPath === '.strains') {
-        return 'A default strain is missing.';
-      }
-      if (keyword === 'required') {
-        return `${prettyname(dataPath, parentSchema)} ${message}`;
-      }
-      if (keyword === 'oneOf' && dataPath.startsWith('.strains')) {
-        return `${prettyname(dataPath, parentSchema)} must be either a Runtime Strain or a Proxy Strain`;
-      }
-      return `${prettyname(dataPath, parentSchema)} ${message}: ${keyword}(${JSON.stringify(data)}, ${JSON.stringify(params)})`;
-    }).join('\n');
-    super(`Invalid configuration:
-${detail}
-
-${msg}`);
-    this._errors = errors;
-  }
-}
 
 class ConfigValidator {
   constructor() {
